% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_intensities.R
\name{infer_intensities}
\alias{infer_intensities}
\title{Infer a matrix of non-negative intensities in NMF}
\usage{
infer_intensities(C, X, esign = "pos", n.cores = 1)
}
\arguments{
\item{C}{Numeric matrices.}

\item{X}{Numeric matrices.}

\item{esign}{A character. Keep elements of matrix \code{D} non-negative ("pos") or not ("all). (default="pos")}

\item{n.cores}{An integer. Number of cores to use. (default=1)}
}
\value{
Fitted matrix \code{D}.
}
\description{
\code{infer_intensities} estimates a non-negative matrix \code{D} that optimizes the objective function \eqn{F = ||X - C*D||^2}
using per-row quadratic programming.
}
