% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_simplex.R
\name{projection_onto_simplex}
\alias{projection_onto_simplex}
\title{Project vector onto a probabilistic simplex.}
\usage{
projection_onto_simplex(unproj, bound)
}
\arguments{
\item{unproj}{A numeric vector. An unprojected vector}

\item{bound}{A numeric. Sum of projected vector elements.}
}
\value{
A projected vector.
}
\description{
\code{projection_onto_simplex} projects a vector \code{unproj} onto a probabilistic simplex of sum \code{bound}.
}
