% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_factorization.R
\name{sim_factors}
\alias{sim_factors}
\title{Simulate matrices to explores \code{vrnmf}}
\usage{
sim_factors(
  m,
  n,
  r,
  simplex = "col",
  distr = "unif",
  frac.zeros = 0.3,
  condition = TRUE,
  noise = 0
)
}
\arguments{
\item{m, n, r}{Integers. Size of matrices. Matrix \code{C} has a size of \code{m*r}
and matrix \code{D} has a size of \code{r*n}.}

\item{simplex}{A character. Either columns ("col") or rows ("row") of matrix \code{C} are projected onto unit simplex.}

\item{distr}{A character. Distribution to simulate matrix entries: "unif" for uniform and "exp" for exponential distributions.}

\item{frac.zeros}{A numeric. Fraction of zeros in matrix \code{C}. It promotes sufficient scattering of matrix column/row vectors.}

\item{condition}{A boolean. Generate more well-conditioned matrix \code{R}.}

\item{noise}{A numeric. Standard deviation of gaussian noise to add.}
}
\value{
List of simulated matrices:

\code{X.noise}, \code{X} - noisy and original matrix \code{X} to decompose.

\code{C}, \code{R} - factorization matrices.
}
\description{
\code{sim_factors} simulates non-negative factorization matrices \code{C} and \code{D}
under a variaty of conditions to explroe factorization \eqn{X = C*D + noise}.
}
