% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simnmf.R
\docType{data}
\name{simnmf}
\alias{simnmf}
\title{Simulated example of noiseless non-negative matrix factorization.}
\format{
A list of 4 items.
\describe{
  \item{X.noise}{noise matrix}
  \item{X}{original matrix}
  \item{C}{left decomposition matrix}
  \item{D}{right decomposition matrix}
}
}
\usage{
simnmf
}
\description{
A dataset containing original, noisy and decomposition matrices.
}
\keyword{datasets}
