% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simnmf.R
\docType{data}
\name{simnmf}
\alias{simnmf}
\title{Simulated example of noiseless non-negative matrix factorization.}
\format{
An object of class \code{list} of length 4.
}
\usage{
simnmf
}
\description{
A dataset containing original, noisy and decomposition matrices.
}
\keyword{datasets}
